feat: Adds `addNS`
takes a ns and an object and adds the ns an non-enumarable toString
function

BREAKING CHANGE: This changes the internal behavior of functions when
dealing with ns. If using this library to add ns to reducers then you
should be fine. But to be safe this is being released as breaking